chore: rename modal search packages files
